requirements.txt
pandas
yfinance
pandas-ta-openbb
tabulate
plotly